Move to channel-less implementation.
When multiple go routines queue to send into a channel, the order of their entry to the channel is unknowable:

https://groups.google.com/g/golang-nuts/c/PWt4r9b40bc/m/lC59KD5TQCwJ

Therefore, we cannot rely on a channel for ordering, as if we overrun its size, order will be non-deterministic.

This PR moves to using a strict ordering by atomically ordered ticket.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>